Feat: token null exception 추가
